Add TodoContainer.js file. Move CRUD operations from App.js to TodoContainer.js
